ingestor/pyproject.toml
[tool.poetry]
name = "ingestor"
version = "0.1.0"
description = ""
authors = ["Aviad Hadad <aviad@sthuck.info>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.26.0"
python-dotenv = "^1.0.1"
pymongo = "^4.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
